src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/selection-shouldnt-expose-shadow-dom-expected.txt

CONSOLE ERROR: line 27: Uncaught TypeError: container.createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
